pyproject.toml
[project]
name = "MyPrompt"
version = "0.1.0"
description = "Produces an XML LLM prompt from a user natural language request"
requires-python = ">=3.11"
dependencies = [
    "annotated-types==0.7.0",
    "blinker==1.9.0",
    "cachetools==5.5.2",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "colorama==0.4.6",
    "coverage==7.8.0",
    "flask==2.3.3",
    "flask-cors==4.0.0",
    "google-ai-generativelanguage==0.6.1",
    "google-api-core==2.24.2",
    "google-api-python-client==2.169.0",
    "google-auth==2.39.0",
    "google-auth-httplib2==0.2.0",
    "google-generativeai==0.5.0",
    "googleapis-common-protos==1.70.0",
    "grpcio==1.71.0",
    "grpcio-status==1.62.3",
    "httplib2==0.22.0",
    "idna==3.10",
    "iniconfig==2.1.0",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "markupsafe==3.0.2",
    "packaging==25.0",
    "pluggy==1.5.0",
    "proto-plus==1.26.1",
    "protobuf==4.25.7",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pydantic==2.11.4",
    "pydantic-core==2.33.2",
    "pyparsing==3.2.3",
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "pytest-flask>=1.0.0",
    "python-dotenv==1.0.0",
    "requests==2.32.3",
    "rsa==4.9.1",
    "tqdm==4.67.1",
    "typing-extensions==4.13.2",
    "typing-inspection==0.4.0",
    "uritemplate==4.1.1",
    "urllib3==2.4.0",
    "werkzeug==3.1.3",
    "pylance>=0.27.1",
    "pytest-mock>=3.0.0",
]

[tool.pytest.ini_options]
markers = [
    "happy_path: Happy path tests",
    "edge_case: Edge case tests"
]
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "function"

[project.optional-dependencies]
dev = ["black>=23.0.0", "isort>=5.0.0", "mypy>=1.0.0"]
